@@ -5,10 +5,10 @@
 
 namespace dsplib {
 
-struct agc_impl;
+struct AgcImpl;
 
 //Adaptively adjust gain for constant signal level output
-class agc
+class Agc
 {
 public:
     //target_level - target output power level
@@ -16,28 +16,28 @@ public:
     //average_len - length of averaging window
     //t_rise - step size for gain updates (rise)
     //t_fall - step size for gain updates (fall)
-    explicit agc(double target_level = 1, double max_gain = 60.0, int average_len = 100, double t_rise = 0.01,
-                 double t_fall = 0.01);
-
-    ~agc();
+    explicit Agc(real_t target_level = 1, real_t max_gain = 60.0, int average_len = 100, real_t t_rise = 0.01,
+                 real_t t_fall = 0.01);
 
     template<typename T>
-    struct result
+    struct Result
     {
         base_array<T> y;
         arr_real gain;
     };
 
-    result<real_t> process(const arr_real& x);
-    result<cmplx_t> process(const arr_cmplx& x);
+    Result<real_t> process(const arr_real& x);
+    Result<cmplx_t> process(const arr_cmplx& x);
 
     template<typename T>
-    result<T> operator()(const base_array<T>& x) {
+    Result<T> operator()(const base_array<T>& x) {
         return this->process(x);
     }
 
 private:
-    std::unique_ptr<agc_impl> _d;
+    std::shared_ptr<AgcImpl> _d;
 };
+
+using agc [[deprecated]] = Agc;
 
 }   // namespace dsplib
