@@ -9,60 +9,6 @@
 namespace dsplib {
 
 //-------------------------------------------------------------------------------------------------
-fir::fir() {
-    _h = arr_real(1);
-    _h[0] = 1;
-}
-
-//-------------------------------------------------------------------------------------------------
-fir_cmplx::fir_cmplx() {
-    _h = arr_cmplx(1);
-    _h[0] = 1;
-}
-
-//-------------------------------------------------------------------------------------------------
-fir::fir(const arr_real& h) {
-    _h = h;
-    _d = zeros(h.size() - 1);
-}
-
-//-------------------------------------------------------------------------------------------------
-fir_cmplx::fir_cmplx(const arr_cmplx& h) {
-    _h = h;
-    _d = arr_cmplx(zeros(h.size() - 1));
-}
-
-//-------------------------------------------------------------------------------------------------
-arr_real fir::process(const arr_real& s) {
-    auto x = _d | s;
-
-    //convolution
-    arr_real r = fir::conv(x, _h);
-
-    //keep filter delay
-    int nd = _h.size() - 1;
-    int nx = x.size();
-    _d = x.slice((nx - nd), nx);
-
-    return r;
-}
-
-//-------------------------------------------------------------------------------------------------
-arr_cmplx fir_cmplx::process(const arr_cmplx& s) {
-    auto x = _d | s;
-
-    //convolution
-    arr_cmplx r = fir_cmplx::conv(x, _h);
-
-    //keep filter delay
-    int nd = _h.size() - 1;
-    int nx = x.size();
-    _d = x.slice((nx - nd), nx);
-
-    return r;
-}
-
-//-------------------------------------------------------------------------------------------------
 template<class T>
 static void _conv(const T* restrict x, const T* restrict h, T* restrict r, int nh, int nx) {
     const int nr = nx - nh + 1;
@@ -76,24 +22,26 @@ static void _conv(const T* restrict x, const T* restrict h, T* restrict r, int n
 }
 
 //-------------------------------------------------------------------------------------------------
-arr_real fir::conv(const arr_real& x, const arr_real& h) {
+template<>
+arr_real FirFilter<real_t>::conv(const arr_real& x, const arr_real& h) {
     arr_real r(x.size() - h.size() + 1);
     _conv(x.data(), h.data(), r.data(), h.size(), x.size());
     return r;
 }
 
 //-------------------------------------------------------------------------------------------------
-arr_cmplx fir_cmplx::conv(const arr_cmplx& x, const arr_cmplx& h) {
+template<>
+arr_cmplx FirFilter<cmplx_t>::conv(const arr_cmplx& x, const arr_cmplx& h) {
     arr_cmplx r(x.size() - h.size() + 1);
     _conv(x.data(), h.data(), r.data(), h.size(), x.size());
     return r;
 }
 
 //-------------------------------------------------------------------------------------------------
-fir_fft::fir_fft(const arr_cmplx& h) {
+FftFilter::FftFilter(const arr_cmplx& h)
+  : _m{h.size()} {
     int fft_len = 1L << nextpow2(2 * h.size());
     _n = fft_len - h.size();
-    _m = h.size();
     _olap = zeros(_m);
     auto dh = conj(h) | zeros(fft_len - h.size());
     _h = fft(dh);
@@ -101,12 +49,12 @@ fir_fft::fir_fft(const arr_cmplx& h) {
 }
 
 //-------------------------------------------------------------------------------------------------
-fir_fft::fir_fft(const arr_real& h)
-  : fir_fft(arr_cmplx(h)) {
+FftFilter::FftFilter(const arr_real& h)
+  : FftFilter(arr_cmplx(h)) {
 }
 
 //-------------------------------------------------------------------------------------------------
-arr_cmplx fir_fft::process(const arr_cmplx& x) {
+arr_cmplx FftFilter::process(const arr_cmplx& x) {
     if (_n == 0) {
         return x;
     }
@@ -120,13 +68,13 @@ arr_cmplx fir_fft::process(const arr_cmplx& x) {
         if (_nx == _n) {
             const auto yy = fft(_x) * _h;
             auto ry = ifft(yy);
-            for (size_t i = 0; i < _m; i++) {
+            for (int i = 0; i < _m; i++) {
                 ry[i] += _olap[i];
             }
-            for (size_t i = 0; i < _m; i++) {
+            for (int i = 0; i < _m; i++) {
                 _olap[i] = ry[i + _n];
             }
-            for (size_t i = 0; i < _n; i++) {
+            for (int i = 0; i < _n; i++) {
                 pr[i] = ry[i];
             }
             pr += _n;
@@ -137,7 +85,7 @@ arr_cmplx fir_fft::process(const arr_cmplx& x) {
 }
 
 //-------------------------------------------------------------------------------------------------
-arr_real fir_fft::process(const arr_real& x) {
+arr_real FftFilter::process(const arr_real& x) {
     return real(process(arr_cmplx(x)));
 }
 
